Removed flipping board functionality
TODO: flip board button, since last approach wasnt  optimal to handle online games, causing problems to the game itself such as mismatch of board coordinates in both players therefore mismatch of the same inside the database.

Approach: Change only the UI and not the virtualBoard class as it should be taken as only source of truth.